cloud_functions/trigger_bq_scripts/requirements.txt
google-cloud-bigquery==2.31.0
google-cloud-storage==1.44.0